+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.0.1] - 2021-12-14
 
+### Fixed
+
+- Create Droplet timeout
+
+- API.js start failed
+
+## [0.0.1-alpha] - 2021-12-14
+
 ### Added
 
 - Create droplet by templates